requirements.txt
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.3.1
scikit_image==0.17.2
scikit_learn==0.24.2
scipy==1.5.4
segmentation_models_pytorch==0.2.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
opencv-python>=4.5